+++ src/formatters.ts
@@ -1,11 +1,13 @@
 import mdcodeblocks from "./formatters/mdcodeblocks";
 import mdtable from "./formatters/mdtable";
 import ndjson from "./formatters/ndjson";
+import plaintext from "./formatters/plaintext";
 
 export const formatters = [
   mdtable,
   ndjson,
-  mdcodeblocks
+  mdcodeblocks,
+  plaintext
 ]
   .sort((a, b) => a.title.localeCompare(b.title));