+++ src/plugin-info.ts
@@ -1,5 +1,5 @@
 /* File will be overwritten by bin/release.sh! */
 export const PLUGIN_INFO = {
-  "pluginVersion": "2.3.0",
-  "pluginReleasedAt": "2025-11-18T13:38:38+0100"
+  "pluginVersion": "3.0.0",
+  "pluginReleasedAt": "2026-06-06T19:00:00+0200"
 }